+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+* Use `autoreleasepool` when rendering image.
+
 ## 0.0.4
 
 * Add - Resolution Parameter to ImageRenderer.render()