+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## master
 
-* Use `autoreleasepool` when rendering image.
+- Add `current` to Options. We can now modify options globally.
+- Use `autoreleasepool` when rendering image.
 
 ## 0.0.4
 
-* Add - Resolution Parameter to ImageRenderer.render()
-* Change - passing editingStack in PixelEditViewControllerDelegate instead of UIImage
+- Add - Resolution Parameter to ImageRenderer.render()
+- Change - passing editingStack in PixelEditViewControllerDelegate instead of UIImage