+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## master
+## 0.0.7
 
 - Add `current` to Options. We can now modify options globally.
+
+## 0.0.6
+
 - Use `autoreleasepool` when rendering image.
 
 ## 0.0.4